// src/src/types/request.d.ts
interface AppReport {
    objectId: string;
    objectName: string;
    objectOwner?: string;
    description?: string;
    isFavourite: boolean;
    documentationURL?: string;
    applicationURL?: string;
    objectType: string;
}